fix: resolve all 58 findings from full codebase audit
Backend: transaction safety for file deletion, N+1 query fixes,
batch saves, JWT secret validation, idempotency key generation,
push token cleanup, coordinate validation, game status guards,
HTML sanitizer hardening, graceful shutdown, logback config.

Frontend: React Query enabled guards, route-level code splitting,
O(1) Map lookups, accessibility, optimistic updates, type safety.

iOS: memory-safe video handling, SyncEngine seek optimization,
AppState init race fix, location overlap guard, reconnect refresh,
ID-based marker matching, join code validation, QR confirmation.

Android: auth error propagation, polling loop fix, runBlocking
removal, DB indexes, unique camera filenames, failed sync warnings,
game status submission guard, shared marker bitmap utils.

Infra: pinned Docker image versions, JAVA_TOOL_OPTIONS consistency,
nginx rate limit 429, proxy timeouts, server_tokens off, certbot
volume read-only, container restart policies.

See docs/audit-fix-summary.md and docs/audit-decisions.md.

+++ ios-app/dbv-nfc-games/Models/Submission.swift
@@ -80,7 +80,7 @@ struct PlayerNotificationResponse: Codable, Identifiable {
 }
 
 struct UnseenCountResponse: Codable {
-    let count: Int
+    let count: Int64
 }
 
 // MARK: - Pending Media Item
